Move TypeScript index to root
