+++ .gitignore
@@ -1,14 +1,7 @@
 *.iml
 node_modules
-src/**/*.js
-src/**/*.js.map
-src/**/*.d.ts
-lib/**/*.js
-lib/**/*.js.map
-lib/**/*.d.ts
-test/**/*.js
-test/**/*.js.map
-test/**/*.d.ts
+**/*.js
+**/*.js.map
+**/*.d.ts
 .nyc_output
 coverage
-
