+++ package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "clean": "del-cli lib/**/*.js lib/**/*.js.map lib/**/*.d.ts test/**/*.d.ts test/**/*.js test/**/*.js.map .nyc_output",
     "build": "npm run compile",
-    "compile-src": "tsc -p lib",
+    "compile-src": "tsc",
     "compile-test": "tsc -p test",
     "compile": "npm run compile-src && npm run compile-test",
     "lint-ts": "tslint 'lib/**/*.ts' --exclude 'lib/**/*.d.ts' 'test/**/*.ts' --exclude 'test/**/*.d.ts'",
@@ -27,12 +27,12 @@
     "url": "https://github.com/Borewit/token-types"
   },
   "files": [
-    "lib/**/*.js",
-    "lib/**/*.d.ts"
+    "index.js",
+    "index.d.ts"
   ],
   "license": "MIT",
-  "main": "./lib/index.js",
-  "typings": "lib/index",
+  "main": "index.js",
+  "typings": "index",
   "bugs": {
     "url": "https://github.com/Borewit/token-types/issues"
   },
