+++ test/test-ieee-754-floats.ts
@@ -2,7 +2,7 @@
 
 import {} from 'mocha';
 import { assert } from 'chai';
-import * as Token from '../lib';
+import * as Token from '..';
 import * as util from './util';
 
 describe('IEEE 754 floats', () => {
