+++ test/test-uint16.ts
@@ -2,7 +2,7 @@
 
 import {} from 'mocha';
 import {assert} from 'chai';
-import * as Token from '../lib';
+import * as Token from '..';
 import * as util from './util';
 
 describe("Parse 16-bit unsigned integer", () => {
