+++ test/tsconfig.json
@@ -1,4 +1,7 @@
 {
-  "extends": "../tsconfig.json"
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "declaration": false
+  }
 }
